Move tests and format them
